+++ src/configs.ts
@@ -1,7 +1,8 @@
-import { dirname, resolve } from 'node:path'
+import { dirname, relative, resolve } from 'node:path'
 import process from 'node:process'
+import { ConfigArray } from '@eslint/config-array'
+import { configArrayFindFiles } from '@voxpelli/config-array-find-files'
 import { bundleRequire } from 'bundle-require'
-import fg from 'fast-glob'
 import { findUp } from 'find-up'
 import c from 'picocolors'
 import { resolve as resolveModule } from 'mlly'
@@ -104,6 +105,41 @@ export async function readConfig(
   if (!Array.isArray(rawConfigs))
     rawConfigs = [rawConfigs]
 
+  // ESLint applies these default configs to all files
+  // https://github.com/eslint/eslint/blob/21d3766c3f4efd981d3cc294c2c82c8014815e6e/lib/config/default-config.js#L66-L69
+  rawConfigs.unshift(
+    {
+      name: 'eslint/defaults/languages',
+      languageOptions: {
+        sourceType: 'module',
+        ecmaVersion: 'latest',
+        parserOptions: {},
+      },
+      linterOptions: {
+        reportUnusedDisableDirectives: 1,
+      },
+    } as FlatConfigItem,
+    {
+      name: 'eslint/defaults/ignores',
+      ignores: [
+        '**/node_modules/',
+        '.git/',
+      ],
+    } as FlatConfigItem,
+    {
+      name: 'eslint/defaults/files',
+      files: ['**/*.js', '**/*.mjs'],
+    } as FlatConfigItem,
+    {
+      name: 'eslint/defaults/files-cjs',
+      files: ['**/*.cjs'],
+      languageOptions: {
+        sourceType: 'commonjs',
+        ecmaVersion: 'latest',
+      },
+    } as FlatConfigItem,
+  )
+
   const rulesMap = new Map<string, RuleInfo>()
 
   // Try resolve `eslint` module from the same directory as the config file
@@ -173,28 +209,39 @@ export async function readConfig(
   }
 }
 
+const noopSchema = {
+  merge: 'replace',
+  validate() {},
+}
+
+const flatConfigNoopSchema = {
+  settings: noopSchema,
+  linterOptions: noopSchema,
+  language: noopSchema,
+  languageOptions: noopSchema,
+  processor: noopSchema,
+  plugins: noopSchema,
+  rules: noopSchema,
+}
+
 export async function globMatchedFiles(
   basePath: string,
   configs: FlatConfigItem[],
 ): Promise<MatchedFile[]> {
   console.log(MARK_INFO, 'Globing matched files')
-  const files = await fg(
-    configs.flatMap(i => i.files ?? []).filter(i => typeof i === 'string') as string[],
-    {
-      cwd: basePath,
-      onlyFiles: true,
-      ignore: [
-        '**/node_modules/**',
-        '**/dist/**',
-        '**/.git/**',
-        ...configs
-          .filter(i => isIgnoreOnlyConfig(i))
-          .flatMap(i => i.ignores ?? [])
-          .filter(i => typeof i === 'string') as string[],
-      ],
-      deep: 5, // TODO: maybe increase this?
-    },
-  )
+
+  const configArray = new ConfigArray(configs, {
+    basePath,
+    schema: flatConfigNoopSchema,
+  })
+
+  await configArray.normalize()
+
+  const files = await configArrayFindFiles({
+    basePath,
+    configs: configArray,
+  })
+
   files.sort()
 
   const ignoreOnlyConfigs = configs.filter(isIgnoreOnlyConfig)
@@ -210,6 +257,7 @@ export async function globMatchedFiles(
 
   return files
     .map((filepath) => {
+      filepath = relative(basePath, filepath)
       const result = matchFile(filepath, configs, ignoreOnlyConfigs)
       if (!result.configs.length)
         return undefined